chore(nest): move things around to have all at the root of dist after build
