+++ examples/bifrost-starter/apps/backend/src/modules/auth/auth.controller.ts
@@ -1,8 +1,8 @@
 import { Body, Req, Res } from '@nestjs/common';
 import { Request, Response } from 'express';
-import { Environment } from 'src/env.validation';
-import { Controller } from 'decorators/controller';
-import { Post } from 'decorators/httpDecorators';
+import { Environment } from '@root/env.validation';
+import { Controller } from '@decorators/controller';
+import { Post } from '@decorators/httpDecorators';
 import { AuthService } from './auth.service';
 import { Credentials } from './interfaces/credentials.dto';
