+++ examples/bifrost-starter/apps/backend/src/modules/auth/tests/auth.controller.spec.ts
@@ -2,11 +2,11 @@ import * as request from 'supertest';
 import * as faker from 'faker';
 import { INestApplication } from '@nestjs/common';
 import { Test, TestingModule } from '@nestjs/testing';
-import { AppModule } from 'src/app.module';
-import { UserService } from 'modules/user/user.service';
-import { UserFactory } from 'modules/user/user.factory';
+import { AppModule } from '@root/app.module';
+import { UserService } from '@modules/user/user.service';
+import { UserFactory } from '@modules/user/user.factory';
 import { Repository } from 'typeorm';
-import { User } from 'modules/user/user.entity';
+import { User } from '@modules/user/user.entity';
 import { getRepositoryToken } from '@nestjs/typeorm';
 import * as cookieParser from 'cookie-parser';
 import { AuthService } from '../auth.service';
