+++ examples/bifrost-starter/apps/backend/src/modules/user/tests/user.controller.spec.ts
@@ -4,9 +4,9 @@ import { Repository } from 'typeorm';
 import * as faker from 'faker';
 import * as request from 'supertest';
 
-import { AppModule } from 'src/app.module';
+import { AppModule } from '@root/app.module';
 import { INestApplication } from '@nestjs/common';
-import { AuthService } from 'auth/auth.service';
+import { AuthService } from '@auth/auth.service';
 import { User } from '../user.entity';
 import { UserFactory } from '../user.factory';
 import { GetUserDto } from '../interfaces/getUser.dto';
