+++ examples/bifrost-starter/apps/backend/src/tests/app.controller.spec.ts
@@ -1,7 +1,7 @@
 import { INestApplication } from '@nestjs/common';
 import { Test, TestingModule } from '@nestjs/testing';
 import * as request from 'supertest';
-import { AppModule } from '../app.module';
+import { AppModule } from '@root/app.module';
 
 describe('AppController', () => {
   let app: INestApplication;
